--- app/src/main/java/com/example/pokedex/model/Pokemon.kt
@@ -1,9 +1,0 @@
-package com.example.pokedex.model
-
-class Pokemon(numero : Int, nome : String, descricao : String, tipo1 : String, tipo2 : String?) {
-    val numero = numero
-    val nome = nome
-    val descricao = descricao
-    val tipo1 = tipo1
-    val tipo2= tipo2
-}
