+++ app/src/main/java/com/example/pokedex/presentation/MainActivity.kt
@@ -6,12 +6,12 @@ import android.view.View
 import android.widget.*
 import androidx.appcompat.widget.Toolbar
 import androidx.fragment.app.FragmentManager
-import com.example.pokedex.fragment.PokemonListFragment
+import com.example.pokedex.presentation.fragment.PokemonListFragment
 import com.example.pokedex.R
 import com.example.pokedex.config.RetrofitConfig
 import com.example.pokedex.controller.BuscaPokemon
 import com.example.pokedex.controller.MostraPokemon
-import com.example.pokedex.model.Pokemon
+import com.example.pokedex.data.model.Pokemon
 import kotlinx.android.synthetic.main.toolbar.*
 import retrofit2.Call
 import retrofit2.Callback
