+++ app/src/main/java/com/example/pokedex/presentation/adapter/PokemonDialogAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.pokedex.adapter
+package com.example.pokedex.presentation.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.pokedex.R
 import com.example.pokedex.controller.MudaCor
-import com.example.pokedex.model.Pokemon
+import com.example.pokedex.data.model.Pokemon
 import kotlinx.android.synthetic.main.row_dialog_fragment.view.*
 
 class PokemonDialogAdapter(private val pokemon: Pokemon) : RecyclerView.Adapter<PokemonDialogAdapter.PokemonDialogViewHolder>() {
