+++ app/src/main/java/com/example/pokedex/presentation/fragment/PokemonDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.example.pokedex.fragment
+package com.example.pokedex.presentation.fragment
 
 import android.content.Context
 import android.os.Bundle
@@ -9,8 +9,8 @@ import androidx.fragment.app.DialogFragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.pokedex.R
-import com.example.pokedex.adapter.PokemonDialogAdapter
-import com.example.pokedex.model.Pokemon
+import com.example.pokedex.presentation.adapter.PokemonDialogAdapter
+import com.example.pokedex.data.model.Pokemon
 
 class PokemonDialogFragment(pokemon: Pokemon, appContext : Context) : DialogFragment(){
