+++ app/src/main/java/com/example/pokedex/service/ListaPokemonService.kt
@@ -1,6 +1,6 @@
 package com.example.pokedex.service
 
-import com.example.pokedex.model.Pokemon
+import com.example.pokedex.data.model.Pokemon
 import retrofit2.Call
 import retrofit2.http.GET
