+++ app/src/main/res/layout/fragment_pokemon_dialog.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".fragment.PokemonDialogFragment">
+    tools:context=".presentation.fragment.PokemonDialogFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerPokemonList"
