+++ app/src/main/res/layout/fragment_pokemon_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.PokemonListFragment">
+    tools:context=".presentation.fragment.PokemonListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerPokemonList"
